Fix logic bug in find_hosts caching
The single = resulted in the value being assigned. If it's non-nil then
it will also be true, even though the code suggests it was intended as a
condition. In f87a1a83f2a3e3e8cad2c241b56695bd99c85c1a it was changed,
but it looks like that may have been accidental.

Fixes: f87a1a83f2a3 ("Fixes #21025 - support for non-default dns and network views.")